Refactor test world helpers
