+++ secrets.yml
@@ -1,5 +1,5 @@
-PCF_CONJUR_ACCOUNT: myorg
-PCF_CONJUR_APPLIANCE_URL: https://conjur-mlee.itd.conjur.net
+PCF_CONJUR_ACCOUNT: pcf
+PCF_CONJUR_APPLIANCE_URL: https://conjur-pcf.itd.conjur.net
 PCF_CONJUR_USERNAME: !var ci/pcf/conjur-admin-user
 PCF_CONJUR_API_KEY: !var ci/pcf/conjur-admin-key
